Bug 1864682 - Use FirefoxAccount directly
This allows us to calculate the initial state without depending on
AccountOnDisk.  The new calculation takes into account the possibility
that a disconnected account, or an account with auth problems, was
persisted.  Maybe this will fix
https://bugzilla.mozilla.org/show_bug.cgi?id=1864667, let's look at the
metrics once this lands.

There were less intrusive ways to do this, but I think FxaClient is
going to be adding more features in general and wanted to make it
realitively easy to access them in FxaAccountManager. The alternative
would be to add any new functionality to OAuthAccount, but that would be
more work per-feature and I'd rather get rid of OAuthAccount than extend it
(https://bugzilla.mozilla.org/show_bug.cgi?id=1854282)

Updated the tests to handle this, which was mostly easy except for the
tests that depended on StatePersistenceTestableAccount.  My plan is to
remove these tests now, but make sure that we add them back before
moving this code to application-services. From what I can tell, these
tests checked that the system:

- generated the correct side effects
- called persistCallback at the correct times
- Handled network retries and auth checks